fix: consolidate member role comparison logic
