+++ packages/db/src/types/index.ts
@@ -2,4 +2,3 @@ export * from "./ApiAccessToken";
 export * from "./LastModifiedBy";
 export * from "./HistoryTable";
 export * from "./Invite";
-export * from "./MemberRole";
